feat: observability, data integrity, and security logging
- Disable Swagger UI/redoc/openapi.json in production (ENV=production)
- Add structured security logging for login success/failure, registration,
  and admin user creation with IP tracking
- Add Docker healthcheck for backend service
- Change amount column from Float to Numeric(12,2) for financial precision
- Alembic migration 004 for amount type change
- Fix Sheets sync to use txn.uuid instead of txn.id (prevents ID leak)

+++ backend/app/models/transaction.py
@@ -4,7 +4,7 @@ import datetime  # noqa: TC003 - needed at runtime by SQLAlchemy Mapped[]
 import uuid as _uuid
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Date, Float, ForeignKey, Index, Integer, String, Uuid
+from sqlalchemy import Date, ForeignKey, Index, Integer, Numeric, String, Uuid
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -27,7 +27,7 @@ class Transaction(Base):
     )
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"), index=True)
     type: Mapped[str] = mapped_column(String(10))
-    amount: Mapped[float] = mapped_column(Float)
+    amount: Mapped[float] = mapped_column(Numeric(12, 2))
     date: Mapped[datetime.date] = mapped_column(Date, index=True)
     category: Mapped[str] = mapped_column(String(100))
     description: Mapped[str | None] = mapped_column(String(500), nullable=True)
